Now everything compiles on Linux
But I highly doubt that it runs, there's still a lot of things to
do. I'm probably converting everything Windows specific to being
cross platform using boost. We'll how that works! :D

Actually, only 8 tests failed, and they are all connected to
FileMaps. Which is a good omen, as the code for the FileMap is
Windows only!